+++ src/_utils/item/index.unit.tsx
@@ -11,14 +11,18 @@ it('Should be displayed in default state without props', () => {
 })
 
 it('Should accept a custom `className`', () => {
-  const customClassName = 'custom'
-  const wrapper = shallow(<Item className={customClassName} />)
-  expect(wrapper.hasClass(customClassName)).toBe(true)
+  const wrapper = shallow(<Item className="test" />)
+  expect(wrapper.hasClass('test')).toBe(true)
 })
 
 it('Should accept a custom `tag`', () => {
-  const wrapper = shallow(<Item tag="li" />)
+  const wrapper = shallow(<Item tag={<li />} />)
   expect(wrapper.type()).toBe('li')
+})
+
+it('Should accept custom props on a custom `tag`', () => {
+  const wrapper = shallow(<Item tag={<li role="option" />} />)
+  expect(wrapper.prop('role')).toBe('option')
 })
 
 it('Should display a left add-on', () => {
@@ -27,13 +31,13 @@ it('Should display a left add-on', () => {
 })
 
 it('Should display title on the left', () => {
-  const wrapper = mount(<Item leftTitle="Test" />)
-  expect(wrapper.find('.kirk-item-leftText .kirk-text-title').exists()).toBe(true)
+  const wrapper = shallow(<Item leftTitle="Test" />)
+  expect(wrapper.find('.kirk-item-leftText .kirk-item-title').exists()).toBe(true)
 })
 
 it('Should display body text on the left', () => {
-  const wrapper = mount(<Item leftBody="Test" />)
-  expect(wrapper.find('.kirk-item-leftText .kirk-text-body').exists()).toBe(true)
+  const wrapper = shallow(<Item leftBody="Test" />)
+  expect(wrapper.find('.kirk-item-leftText .kirk-item-body').exists()).toBe(true)
 })
 
 it('Should display a right add-on', () => {
@@ -42,16 +46,26 @@ it('Should display a right add-on', () => {
 })
 
 it('Should display title on the right', () => {
-  const wrapper = mount(<Item rightTitle="Test" />)
-  expect(wrapper.find('.kirk-item-rightText .kirk-text-title').exists()).toBe(true)
+  const wrapper = shallow(<Item rightTitle="Test" />)
+  expect(wrapper.find('.kirk-item-rightText .kirk-item-title').exists()).toBe(true)
 })
 
 it('Should display body text on the right', () => {
-  const wrapper = mount(<Item rightBody="Test" />)
-  expect(wrapper.find('.kirk-item-rightText .kirk-text-body').exists()).toBe(true)
+  const wrapper = shallow(<Item rightBody="Test" />)
+  expect(wrapper.find('.kirk-item-rightText .kirk-item-body').exists()).toBe(true)
 })
 
 it('Should display a chevron', () => {
-  const wrapper = shallow(<Item chevron />)
-  expect(wrapper.find(ChevronIcon).exists()).toBe(true)
+  const wrapper = shallow(<Item chevron={<ChevronIcon />} />)
+  expect(wrapper.find('.kirk-item-chevron').exists()).toBe(true)
+})
+
+it('Should not have a highlighted state by default', () => {
+  const wrapper = shallow(<Item />)
+  expect(wrapper.hasClass('kirk-item--highlighted')).toBe(false)
+})
+
+it('Should have a highlighted state', () => {
+  const wrapper = shallow(<Item highlighted />)
+  expect(wrapper.hasClass('kirk-item--highlighted')).toBe(true)
 })