+++ src/_utils/item/style.ts
@@ -1,6 +1,6 @@
 import css from 'styled-jsx/css'
 
-import { space } from '_utils/branding'
+import { color, space } from '_utils/branding'
 
 export default css`
   :global(.kirk-item) {
@@ -12,7 +12,31 @@ export default css`
     align-items: center;
     flex: 1;
     border: 0;
+  }
+
+  :global(.kirk-item.kirk-item--narrow) {
+    padding-top: ${space.m};
+    padding-bottom: ${space.m};
+  }
+
+  :global(a.kirk-item) {
     background: none;
+    text-decoration: none;
+    user-select: none;
+    -webkit-tap-highlight-color: ${color.tapHighlight};
+    -webkit-touch-callout: none;
+  }
+
+  /* Button tag fixes */
+  :global(button.kirk-item) {
+    cursor: pointer;
+    text-align: left;
+    width: 100%;
+    font-family: inherit;
+    outline: none;
+    background-color: transparent;
+    -webkit-tap-highlight-color: ${color.tapHighlight};
+    -webkit-touch-callout: none;
   }
 
   /* Text areas */
@@ -26,13 +50,14 @@ export default css`
   }
 
   /* Text */
-  :global(.kirk-item-title) {
+  :global(.kirk-item-title--withBody) {
     margin-bottom: ${space.s};
   }
 
   /* Addons */
   :global(.kirk-item-leftAddon),
-  :global(.kirk-item-rightAddon) {
+  :global(.kirk-item-rightAddon),
+  :global(kirk-item-chevron) {
     display: inline-flex;
     min-width: 24px;
     align-items: center;
@@ -42,7 +67,13 @@ export default css`
     margin-right: ${space.l};
   }
 
-  :global(.kirk-item-rightAddon) {
+  :global(.kirk-item-rightAddon),
+  :global(kirk-item-chevron) {
     margin-left: ${space.l};
+  }
+
+  /* Highlighted state */
+  :global(.kirk-item--highlighted .kirk-item-title) {
+    color: ${color.primary};
   }
 `