+++ src/autoComplete/autoCompleteListItem.tsx
@@ -1,20 +1,16 @@
 import React, { Component } from 'react'
 import ItemChoice, { ItemChoiceStatus } from 'itemChoice'
-import cc from 'classcat'
 
 export interface AutoCompleteListItemProps {
-  readonly item: AutocompleteItem,
-  readonly select: (item:AutocompleteItem) => void,
-  readonly children: React.ReactElement<any>,
-  readonly className?: Classcat.Class,
-  readonly highlighted?: boolean,
-  readonly status?: ItemChoiceStatus,
-  readonly onDoneAnimationEnd?: () => void,
+  readonly item: AutocompleteItem
+  readonly select: (item: AutocompleteItem) => void
+  readonly className?: Classcat.Class
+  readonly status?: ItemChoiceStatus
+  readonly onDoneAnimationEnd?: () => void
 }
 
-export default class AutoCompleteListItem extends Component <AutoCompleteListItemProps> {
+export default class AutoCompleteListItem extends Component<AutoCompleteListItemProps> {
   static defaultProps: Partial<AutoCompleteListItemProps> = {
-    highlighted: false,
     status: ItemChoice.STATUS.DEFAULT,
   }
 
@@ -27,14 +23,16 @@ export default class AutoCompleteListItem extends Component <AutoCompleteListIte
   render() {
     return (
       <ItemChoice
+        href={<button />}
+        narrow
         status={this.props.status}
-        selected={this.props.highlighted}
-        className={cc(['kirk-autoComplete-item', this.props.className])}
+        className={this.props.className}
+        label={this.props.item.title}
+        subLabel={this.props.item.description}
+        leftAddon={this.props.item.icon}
         onMouseDown={this.onMouseDown}
         onDoneAnimationEnd={this.props.onDoneAnimationEnd}
-      >
-        {this.props.children}
-      </ItemChoice>
+      />
     )
   }
 }