+++ src/autoComplete/autoCompleteListStyle.ts
@@ -10,25 +10,26 @@ export default css`
   }
 
   :global(.kirk-autoComplete-item) {
-    padding: ${space.m} ${space.xl};
-  }
-
-  :global(.kirk-autoComplete-primaryText) {
-    color: ${color.primaryText};
-    font-size: ${font.m.size};
-    line-height: ${font.m.lineHeight};
-  }
-
-  :global(.kirk-autoComplete-secondaryText) {
-    color: ${color.secondaryText};
-    font-size: ${font.base.size};
-    line-height: ${font.base.lineHeight};
-    margin-top: ${space.s};
+    position: relative;
+    padding: 0 ${space.xl};
   }
 
   :global(.kirk-autoComplete-item:hover),
   :global(.kirk-autoComplete-item[aria-selected="true"]) {
     background-color: ${color.lightBackground};
     border-radius: ${radius.l};
+  }
+
+  :global(.kirk-autoComplete-item + .kirk-autoComplete-item) {
+    margin-top: ${space.l};
+  }
+
+  :global(.kirk-autoComplete-item + .kirk-autoComplete-item::before) {
+    content : '';
+    position: absolute;
+    top: -${space.m};
+    right: 0;
+    left: 0;
+    border-bottom: 1px solid ${color.border};
   }
 `