+++ src/autoComplete/style.ts
@@ -1,4 +1,5 @@
 import css from 'styled-jsx/css'
+import { color } from '_utils/branding'
 
 export default css`
   .kirk-autoComplete {