+++ src/itemChoice/index.tsx
@@ -1,11 +1,11 @@
-import React, { PureComponent } from 'react'
+import React, { PureComponent, Fragment } from 'react'
 import cc from 'classcat'
-import isEmpty from 'lodash.isempty'
 
-import prefix from '_utils'
-import ChevronIcon from 'icon/chevronIcon'
 import Loader from 'loader'
 import style from './style'
+
+import Item from '_utils/item'
+import ChevronIcon from 'icon/chevronIcon'
 
 export enum ItemChoiceStatus {
   DEFAULT = 'default',
@@ -13,123 +13,106 @@ export enum ItemChoiceStatus {
   CHECKED = 'checked',
 }
 
-interface TypeProps {
-  readonly className?: Classcat.Class,
-  readonly href?: string,
-  readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void,
-  readonly onBlur?: (event: React.FocusEventHandler<HTMLElement>) => void,
-  readonly onFocus?: (event: React.FocusEventHandler<HTMLElement>) => void,
-  readonly onMouseDown?: (event: React.MouseEvent<HTMLElement>) => void,
-}
-
 export interface ItemChoiceProps {
-  readonly className?: Classcat.Class,
-  readonly href?: string | JSX.Element,
-  readonly label?: string,
-  readonly subLabel?: string,
-  readonly children?: React.ReactNode,
-  readonly leftAddon?: React.ReactNode,
-  readonly rightAddon?: React.ReactNode,
-  readonly highlighted?: boolean,
-  readonly selected?: boolean,
-  readonly status?: ItemChoiceStatus,
-  readonly onDoneAnimationEnd?: () => void,
-  readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void,
-  readonly onBlur?: (event: React.FocusEventHandler<HTMLElement>) => void,
-  readonly onFocus?: (event: React.FocusEventHandler<HTMLElement>) => void,
-  readonly onMouseDown?: (event: React.MouseEvent<HTMLElement>) => void,
+  readonly className?: Classcat.Class
+  readonly href: string | JSX.Element
+  readonly label?: string
+  readonly subLabel?: string
+  readonly highlighted?: boolean
+  readonly narrow?: boolean
+  readonly selected?: boolean
+  readonly disabled?: boolean
+  readonly status?: ItemChoiceStatus
+  readonly leftAddon?: React.ReactNode
+  readonly rightAddon?: React.ReactNode
+  readonly onDoneAnimationEnd?: () => void
+  readonly onClick?: (event: React.MouseEvent<HTMLElement>) => void
+  readonly onBlur?: (event: React.FocusEventHandler<HTMLElement>) => void
+  readonly onFocus?: (event: React.FocusEventHandler<HTMLElement>) => void
+  readonly onMouseDown?: (event: React.MouseEvent<HTMLElement>) => void
 }
 
-class ItemChoice extends PureComponent <ItemChoiceProps> {
+class ItemChoice extends PureComponent<ItemChoiceProps> {
   static defaultProps: Partial<ItemChoiceProps> = {
+    disabled: false,
     highlighted: false,
     selected: false,
+    narrow: false,
     status: ItemChoiceStatus.DEFAULT,
     href: '',
   }
+
   static STATUS = ItemChoiceStatus
+
   render() {
     const {
-      children, className, highlighted, selected, status,
-      onClick, onBlur, onFocus, onMouseDown, href, label, subLabel, leftAddon, rightAddon,
+      className,
+      highlighted,
+      narrow,
+      selected,
+      disabled,
+      status,
+      onClick,
+      onBlur,
+      onFocus,
+      onMouseDown,
+      href,
+      label,
+      subLabel,
+      leftAddon,
+      rightAddon,
       onDoneAnimationEnd,
     } = this.props
-    const classNames = cc([prefix({
-      itemChoice: true,
-      'itemChoice--highlighted': highlighted,
-      'itemChoice--withSubLabel': !!subLabel,
-    }), className])
 
-    let rightIcon = <ChevronIcon className={cc(prefix({ chevron: true }))} />
+    const classNames = cc([
+      {
+        'kirk-itemChoice': true,
+        'kirk-itemChoice--disabled': disabled,
+      },
+      className,
+    ])
 
-    if (status === ItemChoiceStatus.LOADING) {
-      rightIcon = <Loader
-        className={cc(prefix({ chevron: true }))}
-        size={24}
-        onDoneAnimationEnd={onDoneAnimationEnd}
-        inline
-      />
-    }
-
-    if (status === ItemChoiceStatus.CHECKED) {
-      rightIcon = <Loader
-        className={cc(prefix({ chevron: true }))}
-        size={24}
-        onDoneAnimationEnd={onDoneAnimationEnd}
-        done
-        inline
-      />
-    }
+    const chevronDisplay =
+      status === ItemChoiceStatus.DEFAULT ? (
+        <ChevronIcon />
+      ) : (
+        <Loader
+          size={24}
+          onDoneAnimationEnd={onDoneAnimationEnd}
+          inline
+          done={status === ItemChoiceStatus.CHECKED}
+        />
+      )
 
     let Component: tag
-    let typeProps: TypeProps
+    const listenerProps: Partial<ItemChoiceProps> = { onClick, onFocus, onBlur, onMouseDown }
+    let componentProps: Partial<ItemChoiceProps> = {}
 
     // If we pass a component to href, we get component type and we merge props
     if (typeof href !== 'string') {
       Component = href.type
-      typeProps = {
-        ...href.props,
-        onClick,
-        onFocus,
-        onBlur,
-        onMouseDown,
-        className: cc([href.props.className, classNames]),
-      }
-    } else if (!isEmpty(href)) {
-      Component = 'a'
-      typeProps = { onClick, onFocus, onBlur, onMouseDown, href, className: classNames }
+      componentProps = { ...listenerProps, ...href.props }
     } else {
-      Component = 'li'
-      typeProps = { onClick, onFocus, onBlur, onMouseDown, className: classNames }
+      Component = 'a'
+      componentProps = { ...listenerProps, href }
     }
 
     return (
-      <Component
-        role="option"
-        aria-selected={selected}
-        {...typeProps}
-      >
-        { leftAddon && (
-          <div className={cc(prefix({ 'itemChoice-leftAddon': true }))}>{leftAddon}</div>
-          )
-        }
-        {label && (
-          <div>
-            <div className={cc(prefix({ 'itemChoice-label': true }))}>{label}</div>
-            {subLabel && (
-              <div className={cc(prefix({ 'itemChoice-subLabel': true }))}>{subLabel}</div>
-            )}
-          </div>
-        )}
-        {children}
-        <div className="kirk-itemChoice-right">
-          {rightAddon && (
-            <div className={cc(prefix({ 'itemChoice-rightAddon': true }))}>{rightAddon}</div>
-          )}
-          {rightIcon}
-        </div>
+      <Fragment>
+        <Item
+          className={classNames}
+          chevron={chevronDisplay}
+          aria-selected={selected}
+          highlighted={highlighted}
+          narrow={narrow}
+          leftAddon={leftAddon ? leftAddon : null}
+          leftTitle={label ? label : null}
+          leftBody={subLabel ? subLabel : null}
+          rightAddon={rightAddon}
+          tag={disabled ? <div aria-disabled="true" /> : <Component {...componentProps} />}
+        />
         <style jsx>{style}</style>
-      </Component>
+      </Fragment>
     )
   }
 }