+++ src/itemChoice/story.tsx
@@ -15,14 +15,13 @@ stories.add(
   withInfo('')(() => (
     <div>
       <ItemChoice
+        href={<button type="button" />}
         label="Choice 1"
         highlighted={boolean('highlighted', false)}
-        selected={boolean('selected', false)}
+        disabled={boolean('disabled', false)}
         onClick={action('onClick')}
       />
-      <ItemChoice
-        label="Choice 2 – A title that is so long it takes 2 lines of text"
-      />
+      <ItemChoice href="" label="Choice 2 – A title that is so long it takes 2 lines of text" />
     </div>
   )),
 )
@@ -31,11 +30,9 @@ stories.add(
   'With loading state',
   withInfo('')(() => (
     <div>
+      <ItemChoice href="" label="Choice 1" status={ItemChoice.STATUS.LOADING} />
       <ItemChoice
-        label="Choice 1"
-        status={ItemChoice.STATUS.LOADING}
-      />
-      <ItemChoice
+        href=""
         label="Choice 2"
         status={ItemChoice.STATUS.CHECKED}
         onDoneAnimationEnd={() => action('animation done')}
@@ -48,9 +45,10 @@ stories.add(
   'With secondary info',
   withInfo('')(() => (
     <div>
-      <ItemChoice label="Choice 1" subLabel="Secondary info" />
-      <ItemChoice label="Choice 2" subLabel="Secondary info" />
+      <ItemChoice href="" label="Choice 1" subLabel="Secondary info" />
+      <ItemChoice href="" label="Choice 2" subLabel="Secondary info" />
       <ItemChoice
+        href=""
         label="Choice 3"
         subLabel="Secondary info – A secondary info that is so long it takes 2 lines of text"
       />
@@ -62,9 +60,9 @@ stories.add(
   'With recommended choice',
   withInfo('')(() => (
     <div>
-      <ItemChoice label="Choice 1" subLabel="Secondary info" />
-      <ItemChoice label="Choice 2" subLabel="Secondary info" highlighted />
-      <ItemChoice label="Choice 3" subLabel="Secondary info" />
+      <ItemChoice href="" label="Choice 1" subLabel="Secondary info" />
+      <ItemChoice href="" label="Choice 2" subLabel="Secondary info" highlighted />
+      <ItemChoice href="" label="Choice 3" subLabel="Secondary info" />
     </div>
   )),
 )
@@ -73,16 +71,8 @@ stories.add(
   'With left addon',
   withInfo('')(() => (
     <div>
-      <ItemChoice
-        label="Choice 1"
-        subLabel="Secondary info"
-        leftAddon={<StarIcon />}
-      />
-      <ItemChoice
-        label="Choice 1"
-        subLabel="Secondary info"
-        leftAddon={<StarIcon />}
-      />
+      <ItemChoice href="" label="Choice 1" subLabel="Secondary info" leftAddon={<StarIcon />} />
+      <ItemChoice href="" label="Choice 2" subLabel="Secondary info" leftAddon={<StarIcon />} />
     </div>
   )),
 )
@@ -91,16 +81,8 @@ stories.add(
   'With right addon',
   withInfo('')(() => (
     <div>
-      <ItemChoice
-        label="Choice 1"
-        subLabel="Secondary info"
-        rightAddon={<StarIcon />}
-      />
-      <ItemChoice
-        label="Choice 1"
-        subLabel="Secondary info"
-        rightAddon="Info"
-      />
+      <ItemChoice href="" label="Choice 1" subLabel="Secondary info" rightAddon={<StarIcon />} />
+      <ItemChoice href="" label="Choice 2" subLabel="Secondary info" rightAddon="Info" />
     </div>
   )),
 )
@@ -109,30 +91,18 @@ stories.add(
   'With basic link',
   withInfo('')(() => (
     <div>
-      <ItemChoice
-        label="Choice 1"
-        href="#anchor"
-      />
-      <ItemChoice
-        label="Choice 2"
-        href="#anchor"
-      />
+      <ItemChoice label="Choice 1" href="#anchor" />
+      <ItemChoice label="Choice 2" href="#anchor" />
     </div>
   )),
 )
 
 stories.add(
-  'With custom tag component',
+  'With button tag',
   withInfo('')(() => (
     <div>
-      <ItemChoice
-        label="Button tag"
-        href={<button type="button" />}
-      />
-      <ItemChoice
-        label="Link tag"
-        href={<a href="#anchor4" />}
-      />
+      <ItemChoice href={<button type="button" />} label="Button tag 1" />
+      <ItemChoice href={<button type="button" />} label="Button tag 2" />
     </div>
   )),
 )